Change list teams API to non-admin specific
